+++ framework/src/main/java/org/checkerframework/framework/type/poly/AbstractQualifierPolymorphism.java
@@ -111,8 +111,8 @@ public abstract class AbstractQualifierPolymorphism implements QualifierPolymorp
               for (Map.Entry<AnnotationMirror, AnnotationMirror> entry : polyQuals.entrySet()) {
                 AnnotationMirror poly = entry.getKey();
                 AnnotationMirror top = entry.getValue();
-                if (type.hasAnnotation(poly)) {
-                  type.removeAnnotation(poly);
+                if (type.hasPrimaryAnnotation(poly)) {
+                  type.removePrimaryAnnotation(poly);
                   if (type.getKind() != TypeKind.TYPEVAR && type.getKind() != TypeKind.WILDCARD) {
                     // Do not add qualifiers to type variables and
                     // wildcards
@@ -133,7 +133,7 @@ public abstract class AbstractQualifierPolymorphism implements QualifierPolymorp
     this.polyScanner =
         new SimpleAnnotatedTypeScanner<>(
             (type, notused) -> {
-              for (AnnotationMirror a : type.getAnnotations()) {
+              for (AnnotationMirror a : type.getPrimaryAnnotations()) {
                 if (qualHierarchy.isPolymorphicQualifier(a)) {
                   return true;
                 }
@@ -253,7 +253,7 @@ public abstract class AbstractQualifierPolymorphism implements QualifierPolymorp
     AnnotationMirrorMap<AnnotationMirror> matchingMapping = new AnnotationMirrorMap<>();
     polyQuals.forEach(
         (polyAnnotation, topAnno) -> {
-          AnnotationMirror annoOnOwner = owner.getAnnotationInHierarchy(topAnno);
+          AnnotationMirror annoOnOwner = owner.getPrimaryAnnotationInHierarchy(topAnno);
           if (annoOnOwner != null) {
             matchingMapping.put(polyAnnotation, annoOnOwner);
           }
@@ -334,8 +334,8 @@ public abstract class AbstractQualifierPolymorphism implements QualifierPolymorp
     for (Map.Entry<AnnotationMirror, AnnotationMirror> kv : polyQuals.entrySet()) {
       AnnotationMirror top = kv.getValue();
       AnnotationMirror poly = kv.getKey();
-      if (polyType.hasAnnotation(poly)) {
-        AnnotationMirror typeQual = type.getAnnotationInHierarchy(top);
+      if (polyType.hasPrimaryAnnotation(poly)) {
+        AnnotationMirror typeQual = type.getPrimaryAnnotationInHierarchy(top);
         if (typeQual != null) {
           if (atypeFactory.hasQualifierParameterInHierarchy(type, top)) {
             polyInstantiationForQualifierParameter.put(poly, typeQual);