+++ framework/src/main/java/org/checkerframework/framework/type/treeannotator/PropagationTreeAnnotator.java
@@ -92,7 +92,7 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
         prev = (prev == null) ? annos : qualHierarchy.leastUpperBounds(prev, annos);
       }
     } else {
-      prev = componentType.getAnnotations();
+      prev = componentType.getPrimaryAnnotations();
     }
 
     assert prev != null
@@ -157,9 +157,9 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
       // TODO: better solution?
       boolean prevIsSubtype = true;
       for (AnnotationMirror am : prev) {
-        if (contextComponentType.isAnnotatedInHierarchy(am)
+        if (contextComponentType.hasPrimaryAnnotationInHierarchy(am)
             && !this.qualHierarchy.isSubtype(
-                am, contextComponentType.getAnnotationInHierarchy(am))) {
+                am, contextComponentType.getPrimaryAnnotationInHierarchy(am))) {
           prevIsSubtype = false;
         }
       }
@@ -167,8 +167,8 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
       // It fails for array initializer expressions. Those should be handled nicer.
       if (contextComponentType.getKind() == componentType.getKind()
           && (prev.isEmpty()
-              || (!contextComponentType.getAnnotations().isEmpty() && prevIsSubtype))) {
-        post = contextComponentType.getAnnotations();
+              || (!contextComponentType.getPrimaryAnnotations().isEmpty() && prevIsSubtype))) {
+        post = contextComponentType.getPrimaryAnnotations();
       } else {
         // The type of the array initializers is incompatible with the context type!
         // Somebody else will complain.
@@ -234,7 +234,7 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
     }
 
     AnnotatedTypeMirror exp = atypeFactory.getAnnotatedType(tree.getExpression());
-    type.addMissingAnnotations(exp.getAnnotations());
+    type.addMissingAnnotations(exp.getPrimaryAnnotations());
     return null;
   }
 
@@ -264,7 +264,7 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
     if (type.getKind() == TypeKind.TYPEVAR) {
       if (exprType.getKind() == TypeKind.TYPEVAR) {
         // If both types are type variables, take the direct annotations.
-        type.addMissingAnnotations(exprType.getAnnotations());
+        type.addMissingAnnotations(exprType.getPrimaryAnnotations());
       }
       // else do nothing.
     } else {